requirements.txt
arabic-reshaper==3.0.0
Faker==36.1.1
jalali_core==1.0.0
jdatetime==5.2.0
numpy==2.2.3
opencv-python==4.11.0.86
pillow==11.1.0
python-bidi==0.6.6
